+++ test/controllers/audits.test.js
@@ -639,6 +639,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -678,6 +679,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -718,6 +720,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -761,6 +764,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -844,6 +848,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -888,6 +893,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -930,6 +936,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -971,6 +978,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -1011,6 +1019,7 @@ describe('Audits Controller', () => {
         getCdnLogsConfig: () => ({}),
         getLlmoConfig: () => ({}),
         getTokowakaConfig: () => ({}),
+        getEdgeOptimizeConfig: () => undefined,
         getBrandProfile: () => ({}),
       });
 
@@ -1077,6 +1086,7 @@ describe('Audits Controller', () => {
           getCdnLogsConfig: () => ({}),
           getLlmoConfig: () => ({}),
           getTokowakaConfig: () => ({}),
+          getEdgeOptimizeConfig: () => undefined,
           getBrandProfile: () => ({}),
         };